+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
 Changelog
 =========
 
+0.0.1.dev2 (2018-05-01)
+-----------------------
+
+Changes
+~~~~~~~
+
+- Removed `http` and `auth` clients in favor of using those in `gordon-gcp`.
+- Renamed configuration key from `zone` to `dns_zone`.
+- Move instance parsing logic from client to authority plugin.
+
+
 0.0.1.dev1 (2018-03-30)
 -----------------------